+++ client/src/commands/resource.show.js
@@ -5,18 +5,28 @@ module.exports = async function command(client) {
     try {
 
         const providerOptions = await client.sendRequest('provider.list').then(
-            result => result.map(data => ({
-                label: data.name,
-                description: data.identifier
-            }))
+            result => result.map(data => {
+                const option = {
+                    identifier: data.identifier,
+                    label: data.name,
+                    description: data.identifier
+                }
+
+                if (data.fromConfiguration) {
+                    option.description = `${data.identifier} - (required_provider in Configuration)`
+                }
+
+                return option
+            })
         )
         const selected = await vscode.window.showQuickPick(providerOptions, {
             title: "Choose a Provider"
         })
         if (null == selected) return
 
+
         const sortOrder = ['overview', 'resources', 'data-sources', 'guides']
-        const resourceOptions = await client.sendRequest('provider.info', selected.description).then(
+        const resourceOptions = await client.sendRequest('provider.info', selected.identifier).then(
             data => data.docs.map(resource => ({
                 ...resource,
                 label: resource.title,
@@ -32,6 +42,7 @@ module.exports = async function command(client) {
         })
         if (null == resource) return
 
+        
         await vscode.env.openExternal(vscode.Uri.parse(resource.docsUrl))
 
     } catch (exception) {