+++ server/src/utility/registry.js
@@ -14,6 +14,11 @@ class Registry {
         return Registry.#instance
     }
 
+    // Found via analyzing terraform files
+    static requiredProvidersAtPath = {}
+
+    // Correlates to the settings
+    static ignoreVersion = false
     static #additionalProviders = []
     static get additionalProviders() {
         return Registry.#additionalProviders
@@ -27,6 +32,7 @@ class Registry {
     }
 
 
+    // Datatypes for getting resource or data-source
     static TYPES = {
         data: 'data-sources',
         resource: 'resources'
@@ -89,11 +95,50 @@ class Registry {
     async getProvidersFromJson(path = `${__dirname}\\data\\providers.json`) {
 
         if (path in this.#cache)
-            return Registry.additionalProviders.map(v => v).concat(this.#cache[path])
+            return Registry.additionalProviders
+                .map(provider => ({ ...provider, fromSettings: true }))
+                .concat(this.#cache[path])
 
         this.#cache[path] = JSON.parse(fs.readFileSync(path))
+            .map(provider => ({
+                ...provider, officialPartnerStatus: true,
+            }))
 
         return await this.getProvidersFromJson(path)
+
+    }
+
+    async getProvidersInConfiguration() {
+
+        const alreadyDefined = {}
+        let providersList = await this.getProvidersFromJson()
+        providersList.forEach(provider => alreadyDefined[provider.identifier] = provider)
+
+        const providersInConfiguration = {}
+        for (const [fullPath, requiredProviders] of Object.entries(Registry.requiredProvidersAtPath)) {
+
+            for (const [name, data] of Object.entries(requiredProviders)) {
+                if (data.source in providersInConfiguration && fullPath.split('\\').length >= providersInConfiguration[data.source].segments) {
+                    continue
+                }
+
+                providersInConfiguration[data.source] = {
+                    name: name,
+                    namespace: data.source.split('/')[0],
+                    identifier: data.source,
+                    version: data.version ?? providersInConfiguration[data.source]?.version,
+                    fsPath: fullPath,
+                    fromConfiguration: true,
+                    segments: fullPath.split('\\').length,
+                    fromSettings: alreadyDefined[data.source]?.fromSettings ?? false,
+                    officialPartnerStatus: alreadyDefined[data.source]?.officialPartnerStatus ?? false,
+                }
+            }
+
+        }
+
+        providersList = providersList.filter(provider => !(provider.identifier in providersInConfiguration))
+        return Object.values(providersInConfiguration).concat(providersList)
 
     }
 
@@ -149,7 +194,7 @@ class Registry {
     async getProviderInfo(identifier, version) {
 
         const docsUrl = `https://${Registry.#endpoint}/providers/{{namespace}}/{{provider}}/{{version}}/docs`
-        const endpoint = null != version ? `v1/providers/${identifier}/${version}` : `v1/providers/${identifier}`
+        const endpoint = null != version && !Registry.ignoreVersion ? `v1/providers/${identifier}/${version}` : `v1/providers/${identifier}`
         const providerInfo = await this.get(endpoint, 'provider', 12 * 60 * 60)
 
         if (null == providerInfo || providerInfo.errors?.at(0)?.toLowerCase() == 'not found') {